automotive-rd/rd1ae: Add support to do device discovery
Currently, The ECAM and MMIO (L and H) regions available on platform
are divided equally between all of root ports and their configuration
values are hardcoded.

This is the initial patch to enable dynamic calculation of PCIe ECAM
and MMIO (L and H) regions and configuration of the same in CMN.
This adds the support to enumerate and calculate the MMIO carveouts
size requirement for all the devices present under a given root port.

This can further reduce the number of devices that can be attached to
the device tree and user may not be able to move the device from one
RP to another RP.

Signed-off-by: Nishant Sharma <nishant.sharma@arm.com>
Signed-off-by: Ziad Elhanafy <ziad.elhanafy@arm.com>