module: add support for atu mmio
The ATU module allows accesses to memory regions that is outside the
processor's address space. The ATU MMIO module extends this
functionality to provide read/write functionality at byte, halfword,
word and double word  granularity. This is achieved by utilizing one
of the ATU translation region to map a memory region that covers the
requested MMIO read/write address and the associated size.

Signed-off-by: Nishant Sharma <nishant.sharma@arm.com>
Change-Id: I437b9f744dd4847c9bb068efa502f57ab115e848