tc3: Correct MPMM configuration
Three changes made in this commit:

- Fix the num_perf_limit structure field to match the number of fields
in the threshold perf array. Without this, the MPMM module will not
select the correct limit.

- Configure the default_perf_limit to match the lowest frequency given
in the threshold perf array. This matches what was done in previous
platforms, ensuring that by default we use the lowest frequency.

- Set the core starts online field to false to core0. The MPMM module
counts online cores fairly simply and simply increments a counter when the
core starts up and decrements when the core goes offline. Core0 boot
also causes the counter to increment leading the module to assume that
3 little cores are online when both are booted. This prevents the
threshold bitmaps being calculated correctly.

Signed-off-by: Jackson Cooper-Driver <jackson.cooper-driver@arm.com>