automotive-rd/rd1ae: Add Transport and MHUv3 module configurations
Two transport channels are required for supporting platform boot -
channel for communicating PSCI messages with the application core,
channel for requesting the secure platform agent firmware to configure
the ATU. Add these as configuration data for the transport module.

MHUv3 module configurations:

Two instances of the MHUv3 doorbell channel are being initially added as
configuration data for the MHUv3 module. The first doorbell channel is
used to send messages between AP and SCP while the second doorbell
channel used between RSE and SCP. The configuration data for these
channels include the interrupt number and the base addresses of the
corresponding Postbox and Mailbox.

Also add the AP Trusted SRAM region to the list of MPU regions.

Signed-off-by: Luca Fancellu <luca.fancellu@arm.com>
Signed-off-by: Ziad Elhanafy <ziad.elhanafy@arm.com>
Signed-off-by: Divin Raj <divin.raj@arm.com>