automotive-rd/rd1ae: Let modules subscribe to platform notifications
Three modules are dependent on the security engine on the platform to
setup the GPC in the control block to be enabled so that these modules
can access addresses outside the SCP subsystem address space. These
three modules are apcontext, sds and transport. So update the config
data of these modules to subscribe to the platform notification.

Signed-off-by: Ziad Elhanafy <ziad.elhanafy@arm.com>