automotive-rd/rd1ae: Add IO block driver
This driver acts as an abstraction layer between common IO Block
programming logic and driver of the underlying hardware. On RD1AE,
IO Block contains NI-Tower, TCU, PCIe and Non PCIe blocks. The NOC S3
is configurable network interconnect unit and connects rest of the blocks
to the CMN. This driver defines IO Block layout for RD1AE platform.
The layout info is passed through the mapping config info and is used by
the module to program a region in NoC.

Enable the support to add static memory mappings in NoC and export APIs
for runtime configuration.

Add a generic interface for IO block drivers.

Add IO block interface header for the IO block driver to consume and
define its APIs.

Signed-off-by: Ziad Elhanafy <ziad.elhanafy@arm.com>