automotive-rd/rd1ae: Enable IO block driver
IO block driver represents the layout in which IO virtualisation block
is laid out on the platform. This driver is needed to configure PCIe
rootport attached to the IO virtualisation block.

Add the configuration files for Tower NCI and IO Block modules and
enable them for scp_ram firmware.

Also enable NoC S3 module as it is a dependency of the IO block module.

Signed-off-by: Ziad Elhanafy <ziad.elhanafy@arm.com>