automotive-rd/rd1ae: Enable IO virtualisation block programming
The IO virtualisation block is made up of the SystemMemory Management
Unit (SMMU), Tower NCI and Generic Interrupt Controller (GIC) IPs. It
provides address translation support for IO devices.

This driver enables the support to configure the block dynamically.
The driver temporarily maps the entire memory carveout to a block and
does PCIe buswalk. During PCIe buswalk, the pcie_enumeration module
checks and cputures the resource requirement of each of the PCIe end
point. The module then uses this data to calculate and allocate
carveout regions for ECAM, MMIOL and MMIOH regions and program/update
the same in NCI and CMN.

Signed-off-by: Ziad Elhanafy <ziad.elhanafy@Arm.com>