scmi_clock: Add SCMI notifications related with clock rate
This patch implements the following as per specified in SCMIv3.1,

Commands:
- CLOCK_RATE_NOTIFY command which allows the agents to request
notification when there is a change in clock rate on a clock device.
- CLOCK_RATE_CHANGE_REQUESTED_NOTIFY command which allows the agents
to request notification when there is a request for change in clock
rate on a clock device from another agent or entity in the system.

Notifications:
- CLOCK_RATE_CHANGED notification which is sent to the subsribing
agents when there is a change in clock rate.
- CLOCK_RATE_CHANGE_REQUESTED notification which is sent to the subsribing
agents when there is a request for change in clock rate.

Update CLOCK_ATTRIBUTES command:
- adds support for clock rate change notification
- adds support for clock rate change requested notification
- adds clock enable delay (currently unsupported)

Signed-off-by: Md Golam Tanvir Zim <mdgolamtanvir.zim@arm.com>
Change-Id: I77b80371c92a3d6ead71ea0c116b7e4ef0d732b2