tc4: Statically initialise PPU_V1 configuration
This configuration can easily be statically configured as nothing about
it needs to be determined dynamically. To facilitate this, add FOR_EACH
macros for the clusters and cores. Use these to initialise the
structure.

We also add a new enum for the PPU elements in a
similar way to other modules. This will be used to make for easier
tracking of the PPU elements (currently they just use immediate values).

Signed-off-by: Jackson Cooper-Driver <jackson.cooper-driver@arm.com>