tc4: Add SYSTOP0/1 to PPU config
Add these two power domains to the PPU_V1 configuration. These are
required to be enabled prior to the use of any features external to the
SCP (including accessing system registers).

Note that default_power_on is set to true which forces the PPU module
to power them on during the initialisation stage.

Signed-off-by: Jackson Cooper-Driver <jackson.cooper-driver@arm.com>