tc4: Add SYSTOP1 to mod_system_power and use PPU
Previously, the underlying driver used by the mod_system_power was the
mock power domain module. This is replaced with the PPU_V1 module as per
the hardware.

SYSTOP1 is also added to the system power module. Note that the
power_domain module does not change as only a single SYSTOP power domain
is exposed to it from the system power module.

Signed-off-by: Jackson Cooper-Driver <jackson.cooper-driver@arm.com>