module/mpmm_v2: Add new Power Management Architecture compliant version of MPMM
This change adds the new version of mpmm which complies with the latest Power
Management Architecture in SCP. It replaces the older model which maps the gear
combinations to DVFS OPPs. The new model assigns weights to gears normalised on
maximum allowed power. It compensates for offline cores and evaluates the safest
and most efficient power limit to avid MPMM throttling.

Signed-off-by: Abhishek Rawat <abhishek.rawat@arm.com>