arch: Introduce Aarch64 architecture support
This commit introduces support for a subset of the AArch64
architecture required to boot on Armv8-R64 devices.

SCP-firmware on AArch64 will (for now) run exclusively at the initial
exception level, which is S-EL2 on Armv8-R64.

The port includes:

 * A crt0 which initializes registers at S-EL2 and installs an initial
   PMSA memory map
 * An exception table
 * Support for both GCC and Clang, but not ArmClang
 * libc stubs (the equivalents in libgloss' nosys generate warnings)
 * Some architectural helper macros, in the absence of CMSIS support for
   AArch64.

Interrupt support using the GIC will be added later - the related
functions are currently stubs.

Signed-off-by: Peter Hoyes <peter.hoyes@arm.com>