arch/common: Fix Clang linker section syntax
The .eh_frame and .ARM.exidx, which are used for stack unrolling data,
are required by the lld linker, which is currently used for
arm-none-eabi but not aarch64-none-elf.

lld seems to parse output sections without an object name (e.g.
"(.eh_frame)" successfully, but ld requires the object name to be
explicit (e.g. "*(.eh_frame)").

Add asterixes to the output section definitions so that linking is
successfully with Clang compilation regardless of the underlying linker
used.

Signed-off-by: Peter Hoyes <peter.hoyes@arm.com>