armv8-a: Use arch_libc_hooks with newlib
Newlib's archive depends on symbols defined in either libgloss
(which throw warnings with "libnosys") or elsewhere, otherwise
"undefined reference" errors are thrown with with Arm GCC >= 11.2.

Use the newly available arch_libc_hooks.c to provide these symbols.

Because newlib's archive is implicitly passed to the linker prior to the
arch_libc_hooks.c archive, it is necessary to force the linker to
prevent the removal of these symbols from newlib, so that they ar
eventually linked to those defined in a later archive.

Signed-off-by: Peter Hoyes <peter.hoyes@arm.com>