aarch64: Fix exception table
Ensure the exception table is aligned to the next 2K, as required by
VBAR_ELx.

We are using SP_ELX, not SP_EL0 so provide the set of exception vectors
for this too. Provide the remaining set of exception vectors for lower
ELs for completeness too.

Replace some tabs with spaces.

Signed-off-by: Peter Hoyes <peter.hoyes@arm.com>