aarch64: Implement interrupt setup and handling
The AArch64 minimal interrupt support is based on the architectrual
GIC (General Interrupt Controller). It has the following limitations
in order to conform to the framework interrupt API:

 * All interrupts have the same priority (0).
 * NMIs (non-maskable interrupts) are not supported.
 * The extended SPI and PPI ranges are not supported.

In order to configure interrupts, the address of the GIC distributor and
redistributor are required. These are provided by the header file
fmw_gic.h, which must be provided by an AArch64 product.

The GIC and CPU interface initialization will be deferred to a dedicated
module, so that additional initialization can be performed prior to the
primary GIC (e.g. for multiview configuration).

Only Group 0 interrupts are enabled, which are handled by the FIQ
handler at the highest implemented exception level.

Implement the fiq_handler and replace the stubs in arch_gic.c with the
implementation. Also implement the interrupt enable/disable methods in
arch_helpers.h. Update the AArch64 documentation.

Signed-off-by: Peter Hoyes <peter.hoyes@arm.com>