gicx00: Introduce module
Introduce a module to perform the GIC initialization prior to unmasking
interrupts.

It contains specific logic to power on the redistributors of
GIC600/GIC600AE/GIC720AE interrupt controllers.

This logic is performed by a module instead of in the AArch64
architecture to allow for other initialization to happen prior to the
GIC (e.g. for multiview configuration).

Signed-off-by: Peter Hoyes <peter.hoyes@arm.com>