ci: Avoid LOG_LEVEL and PRODUCT env var clashes
The CI configuration sets the env var LOG_LEVEL to "" if no explicit log
level is specified in the job. This means no log level is passed to
check_build.py but this CI env var is still picked up by Makefile.cmake,
causing it to always build with DEBUG logs instead of the default for
the build type (INFO/WARN for debug/release respectively).

Do not export the LOG_LEVEL variable to avoid this clash and use the
defaluts in Makefile.cmake. Similarly for PRODUCT.

Signed-off-by: Peter Hoyes <peter.hoyes@arm.com>